test(security): split taskpane CSP coverage
