+++ package.json
@@ -27,7 +27,7 @@
     "test:models": "node --test --experimental-strip-types tests/model-ordering.test.ts",
     "test:recovery": "node --test --experimental-strip-types --import ./scripts/register-test-ts-loader.mjs tests/recovery-log-persistence.test.ts tests/recovery-log-restore.test.ts tests/recovery-log-format.test.ts tests/recovery-log-structure.test.ts tests/recovery-filtering.test.ts tests/recovery-retention.test.ts",
     "test:context": "node --test --experimental-strip-types --import ./scripts/register-test-ts-loader.mjs tests/tool-result-shaping.test.ts tests/tool-output-truncation.test.ts tests/context-injection.test.ts tests/execution-policy.test.ts tests/trace-dependencies.test.ts tests/explain-formula-logic.test.ts tests/explain-formula-tool.test.ts tests/with-workbook-coordinator.test.ts tests/compaction-thresholds.test.ts tests/auto-compaction.test.ts tests/archived-history.test.ts tests/action-queue.test.ts tests/taskpane-runtime-utils.test.ts tests/session-restore-selection.test.ts tests/tab-layout.test.ts tests/tab-layout-persistence.test.ts tests/session-title.test.ts tests/workbook-context.test.ts tests/instructions-store.test.ts tests/conventions-store.test.ts tests/system-prompt.test.ts tests/tool-capabilities.test.ts tests/tool-disclosure.test.ts tests/skills-frontmatter.test.ts tests/docs-skills-drift.test.ts tests/markdown-preprocess.test.ts tests/skills-external-store.test.ts tests/skills-read-cache.test.ts tests/skills-tool.test.ts tests/builtins-registry.test.ts tests/custom-gateways.test.ts tests/prefix-churn.test.ts tests/experimental-tool-gates.test.ts tests/experimental-command.test.ts tests/tmux-tool.test.ts tests/python-run-tool.test.ts tests/execute-office-js-tool.test.ts tests/libreoffice-convert-tool.test.ts tests/python-transform-range-tool.test.ts tests/cell-diff.test.ts tests/workbook-change-audit.test.ts tests/recovery-log-persistence.test.ts tests/recovery-log-restore.test.ts tests/recovery-log-format.test.ts tests/recovery-log-structure.test.ts tests/workbook-save-boundary-monitor.test.ts tests/manual-full-backup.test.ts tests/files-workspace.test.ts tests/files-tool.test.ts tests/files-dialog-filtering.test.ts tests/files-dialog-status.test.ts tests/extensions-manager-tool.test.ts tests/extensions-runtime-mode.test.ts tests/extensions-permissions.test.ts tests/extensions-runtime-manager-sandbox.test.ts tests/extension-loader.test.ts tests/extension-module-import.test.ts tests/extension-api-capabilities.test.ts tests/runtime-manager-activation-http-auth.test.ts tests/connection-manager.test.ts tests/extensions-storage-store.test.ts tests/extension-overlay-integration.test.ts tests/extensions-hub-extension-connections.test.ts tests/extensions-hub-refresh.test.ts tests/extension-widget-api.test.ts tests/overlay-dialog.test.ts tests/sandbox-runtime-schema.test.ts tests/integrations-store.test.ts tests/external-fetch.test.ts tests/network-utils.test.ts tests/bridge-http-utils.test.ts tests/error-path-matrix.test.ts tests/web-search-config.test.ts tests/web-search-setup-detection.test.ts tests/web-search-tool.test.ts tests/fetch-page-tool.test.ts tests/mcp-tool.test.ts tests/mcp-config.test.ts tests/recovery-filtering.test.ts tests/recovery-retention.test.ts tests/python-pyodide-fallback.test.ts",
-    "test:security": "node --test --experimental-strip-types tests/proxy-target-policy.test.mjs tests/cors-proxy-server.security.test.mjs tests/tmux-bridge-server.test.mjs tests/python-bridge-server.test.mjs tests/extension-source-policy.test.ts tests/marked-safety-policy.test.ts tests/oauth-storage-security.test.ts tests/web-search-provider-hosts.test.mjs tests/vercel-ignore-command.test.mjs",
+    "test:security": "node --test --experimental-strip-types tests/proxy-target-policy.test.mjs tests/cors-proxy-server.security.test.mjs tests/tmux-bridge-server.test.mjs tests/python-bridge-server.test.mjs tests/extension-source-policy.test.ts tests/marked-safety-policy.test.ts tests/oauth-storage-security.test.ts tests/web-search-provider-hosts.test.mjs tests/taskpane-csp.test.mjs tests/vercel-ignore-command.test.mjs",
     "validate": "npx office-addin-manifest validate manifest.xml",
     "sideload": "npx office-addin-debugging start manifest.xml desktop --app excel",
     "prepare": "node scripts/install-githooks.mjs"
