+++ tests/web-search-provider-hosts.test.mjs
@@ -2,12 +2,9 @@ import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 import test from "node:test";
 
+import { readTaskpaneConnectSrcTokens } from "./helpers/taskpane-csp.mjs";
 import { PROXY_REACHABILITY_TARGET_URL } from "../src/auth/proxy-validation.ts";
 import { WEB_SEARCH_PROVIDER_ENDPOINT_HOSTS } from "../src/tools/web-search-config.ts";
-
-function isRecord(value) {
-  return typeof value === "object" && value !== null;
-}
 
 function uniqueHosts(hosts) {
   return Array.from(new Set(hosts));
@@ -28,58 +25,6 @@ async function readProxyAllowlistHosts() {
   );
 
   return new Set(hosts);
-}
-
-async function readTaskpaneCspDirectiveTokens(directiveName) {
-  const raw = await readFile(new URL("../vercel.json", import.meta.url), "utf8");
-  const parsed = JSON.parse(raw);
-  if (!isRecord(parsed)) {
-    throw new Error("Invalid vercel.json root structure");
-  }
-
-  const headersRaw = parsed.headers;
-  if (!Array.isArray(headersRaw)) {
-    throw new Error("vercel.json is missing top-level headers array");
-  }
-
-  const taskpaneEntry = headersRaw.find((entry) => isRecord(entry) && entry.source === "/src/taskpane.html");
-  if (!isRecord(taskpaneEntry)) {
-    throw new Error("vercel.json is missing /src/taskpane.html header configuration");
-  }
-
-  const headerListRaw = taskpaneEntry.headers;
-  if (!Array.isArray(headerListRaw)) {
-    throw new Error("/src/taskpane.html entry has no headers array");
-  }
-
-  const cspEntry = headerListRaw.find((entry) => isRecord(entry) && entry.key === "Content-Security-Policy");
-  if (!isRecord(cspEntry) || typeof cspEntry.value !== "string") {
-    throw new Error("Missing Content-Security-Policy value for /src/taskpane.html");
-  }
-
-  const directive = cspEntry.value
-    .split(";")
-    .map((entry) => entry.trim())
-    .find((entry) => entry.startsWith(`${directiveName} `));
-
-  if (!directive) {
-    throw new Error(`CSP missing ${directiveName} directive`);
-  }
-
-  const tokens = directive
-    .split(/\s+/)
-    .slice(1)
-    .filter((token) => token.length > 0);
-
-  return new Set(tokens);
-}
-
-async function readTaskpaneConnectSrcTokens() {
-  return readTaskpaneCspDirectiveTokens("connect-src");
-}
-
-async function readTaskpaneScriptSrcTokens() {
-  return readTaskpaneCspDirectiveTokens("script-src");
 }
 
 test("proxy default host allowlist includes all web-search provider hosts", async () => {
@@ -109,16 +54,3 @@ test("taskpane CSP connect-src allows all web-search provider hosts", async () =
   }
 });
 
-test("taskpane CSP allows Pyodide CDN host in script-src and connect-src", async () => {
-  const connectTokens = await readTaskpaneConnectSrcTokens();
-  const scriptTokens = await readTaskpaneScriptSrcTokens();
-
-  assert.ok(connectTokens.has("https://cdn.jsdelivr.net"), "Missing jsDelivr in CSP connect-src");
-  assert.ok(scriptTokens.has("https://cdn.jsdelivr.net"), "Missing jsDelivr in CSP script-src");
-});
-
-test("taskpane CSP allows blob module imports in script-src", async () => {
-  const scriptTokens = await readTaskpaneScriptSrcTokens();
-
-  assert.ok(scriptTokens.has("blob:"), "Missing blob: in CSP script-src");
-});
